requirements.txt
selenium-wire
requests
psycopg2-binary
regex
python-dotenv